Refactored ForumPostsViewController implementation by functionality into extensions.
